don't provide feedback function to checkers
This is a breaking change. Rather than passing a feedback function to the checker the new contract is that that the checker should just return an R list which includes "message" and "correct" fields as well as (optionally) "type" and "location" fields.

This is done so that checking packages can create their own R functions for returning feedback (including rich feedback based on e.g. markdown composed by the instructur) and not have to pass the 'feedback' function into all of those helper functions.