Possible fix for #50
I’ve added a check for tutor firing a “section_skipped” event on an
undefined element. I grabbed the source for
https://tutorials.shinyapps.io/01-data-basics/ and ran it locally. I
initially did see the console error after a “section_skipped” event for
an undefined element - but then didn’t see it again, in spite of many
attempts to reproduce it.  Anyway, the check I’ve added won’t hurt, and
may indeed fix the problem.