set up userData$learnr_state
This allows for waiting for a round trip from the browser to make sure information is initialized.  Once initialized, prior question submissions can be retrieved.

Without this, questions will try to retrieve information too early.  Questions should not be displayed until prior submission is retrieved.  Do not want to display a blip of temp state.