Properly enforce timelimits >= 60s
If an exercise had an assigned timelimit of >= 60s, it would never be reaped. This is because by default when you subtract times it becomes a `difftime` object which will be presented in arbitrary units. Around the 60s mark, you can see we switch from:

> str(Sys.time() - start_time)
 'difftime' num 58.2525360584259
 - attr(*, "units")= chr "secs"

to:

> str(Sys.time() - start_time)
 'difftime' num 1.01057458321253
 - attr(*, "units")= chr "mins"

When you compare on this time, it compares 1 to your assigned time limit and permits it to continue computing.

We need to force the time comparison into seconds in order to accurately compare to our time limit.