Validate the ability to create an idb store (#417)
The prior checks (line 1437 currently) test whether or not the browser supports the necessary features, but it doesn't actually exercise whether or not we're allowed to create a store here. Some browsers (such as Safari) have additional restrictions in e.g. cross-origin iframes that preclude us from successfully opening an idb store. This check ensures that we can actually create a store, or fails fast if not.

The error caused issues whether or not idb was ultimately used as the store for learnr state.

Without this, Safari is throwing the following error:

[Error] SecurityError: IDBFactory.open() called in an invalid security context
	open (idb-keyval-iife-compat.min.js:7:843)
	e (idb-keyval-iife-compat.min.js:7:843)
	(anonymous function) (tutorial.js:1518)
	(anonymous function) (tutorial.js:1705)
	c (jquery.min.js:2:28300)
	fireWith (jquery.min.js:2:29041)
	l (jquery.min.js:2:79812)
	(anonymous function) (jquery.min.js:2:82256)
[Error] SyntaxError: JSON Parse error: Unexpected identifier "object"
	r (chunk-vendors.a5db6293.js:83:31379)