Prefer rmarkdown::render() for testing for errors
The errors are still surfaced by render() and the test should be runnable on CRAN.

I included a skip if pandoc isn't available, since if the knit step is successful, pandoc will be called.

This also protects against tests where rmarkdown::run() should fail but doesnt, in which case run() will launch the tutorial and run indefinitely.