Remove focus from editor after setValue() to avoid false diffs from blinking cursor
The setValue() method of exercise inputs triggers a click on
the Run button after updating the editor text. The Run button
sends the input value to Shiny and then puts the focus back on
the editor. We test for window.shinytest and if present remove
focus from the editor with a setTimeout(). Otherwise the cursor
blinking is intermittently picked up by the screenshots.