Update learnr for compatibility with shinytest (#527)
* babelify and prettify tutorial-i18n-init.js

via babeljs.io/repl with target "phantomjs 2.1"

* Remove focus from editor after setValue() to avoid false diffs from blinking cursor

The setValue() method of exercise inputs triggers a click on
the Run button after updating the editor text. The Run button
sends the input value to Shiny and then puts the focus back on
the editor. We test for window.shinytest and if present remove
focus from the editor with a setTimeout(). Otherwise the cursor
blinking is intermittently picked up by the screenshots.

* Move comments

* Remove `typeof_()` helper; Replace `Object.*` methods with jQuery methods.

Co-authored-by: Garrick Aden-Buie <garrick@adenbuie.com>
Co-authored-by: Barret Schloerke <schloerke@gmail.com>