Inline `global_setup` into exercise objects on creation
Previously we tried to store the global setup in a separate shiny prerendered chunk with the name "__setup__".
This change avoids potential exercise id clashes with a user-created chunk labelled "__setup_-".
It also avoids a tricky knitr source hook issue when calling `rmarkdown::run()` interactively that would cause the global setup chunk to be lost.