Don't call `parallel::mccollect()` after collecting the result
When the external evaluator was written ~5 years ago circa R 3.3, it's possible this step was required. But AFAICT from the source of `parallel::mccollect()`, this is no longer necessary. The `timeout` argument applies to _finding_ the forked process; if the process has finished it returns the complete result. Once a process completes, {parallel} expects users **not to collect again**, so our second call to `parallel::mccollect()` appears to be doing nothing more than throwing a warning:

```
Warning message:
In selectChildren(jobs, timeout) :
  cannot wait for child 43193 as it does not exist
```

https://github.com/wch/r-source/blob/8454163eed23ffd2703db55c85fa1f1576288308/src/library/parallel/R/unix/mcparallel.R#L65-L80